Move dict_without and filter_none to utils module
